+++ lib/src/utils/delay.dart
@@ -1,0 +1,3 @@
+/// Creates a future that completed with `null` after a delay.
+Future<void> delay(int milliseconds) =>
+    Future.delayed(Duration(milliseconds: milliseconds));