+++ lib/src/utils/index.dart
@@ -1,2 +1,3 @@
 export 'add_null.dart';
+export 'delay.dart';
 export 'identity.dart';